Social Media App : made some ui changes as well as some code changes, and moved logout to ProfileActivity
